Status - Unable to display status output on Windows source and binary build #4339

+++ hummingbot/core/data_type/LimitOrder.pxd
@@ -23,7 +23,7 @@ cdef extern from "../cpp/LimitOrder.h":
                    PyObject *price,
                    PyObject *quantity,
                    PyObject *filledQuantity,
-                   long creationTimestamp,
+                   long long creationTimestamp,
                    short int status)
         LimitOrder(const LimitOrder &other)
         LimitOrder &operator=(const LimitOrder &other)
